+++ packages/mocha/CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- `latency()` method to control delayed test timeouts
+
 ### Changed
 
 - upgraded `@bigtest/convergence` to `^0.4.0`
+- `it.pause` will always set the timeout to zero
 
 ## [0.2.2] - 2018-03-03